[3주차][기본] 결제 포인트 잔액과 콘서트가격 비교 로직 추가 #20
1. 콘서트 가격 int -> Long 타입 변경
2. 콘서트 가격, 포인트 잔액 비교 유효성 로직 추가 -> 도메인 단위테스트 작성

+++ src/main/java/kr/hhplus/be/server/src/domain/model/Concert.java
@@ -30,7 +30,7 @@ public class Concert {
     private String name;
 
     @Schema(description = "티켓 가격", example = "150000")
-    private int price;
+    private Long price;
 
     @Schema(description = "콘서트 날짜", example = "2025-05-01")
     private String date;
@@ -46,7 +46,7 @@ public class Concert {
     private ConcertSeat concertSeat;
 
     // 새로운 Concert 객체 생성 시, 자동으로 좌석을 50개 생성하여 seat 리스트에 넣어줌
-    public Concert(Long concertId, String name, int price, String date, String time, String location) {
+    public Concert(Long concertId, String name, Long price, String date, String time, String location) {
         this.concertId = concertId;
         this.name = name;
         this.price = price;
